+++ src/pages/mentorPages/MentorPage.tsx
@@ -18,6 +18,8 @@ import {
 
 import { ProfileViewModal } from './components/ProfileViewModal';
 import { ViewBookingsModal } from './components/ViewBookingsModal';
+import { MentorDashboard } from './components/MentorDashboard';
+import { MenteeDashboard } from './components/MenteeDashboard';
 import { default as BookingModal } from '../../components/widgets/MentorAlgorithm/booking/BookingModal';
 import { default as CalComModal } from '../../components/widgets/MentorAlgorithm/CalCom/CalComModal';
 import { default as MentorAvailability } from '../../components/widgets/MentorAlgorithm/MentorAvailability';
@@ -54,10 +56,7 @@ export default function MentorPage() {
 
   const [currentPage, setCurrentPage] = useState(1);
   const [profilesPerPage] = useState(12);
-  const [isBookingWidgetMinimized, setIsBookingWidgetMinimized] = useState(false);
   const [isMenteesSectionMinimized, setIsMenteesSectionMinimized] = useState(false);
-  const [isProfileCardExpanded, setIsProfileCardExpanded] = useState(false);
-  const [isAvailabilityCardExpanded, setIsAvailabilityCardExpanded] = useState(false);
 
   // Custom hooks for data management
   const {
@@ -219,10 +218,6 @@ export default function MentorPage() {
 
   const handleViewAllBookings = () => {
     setShowViewBookingsModal(true);
-  };
-
-  const toggleBookingWidget = () => {
-    setIsBookingWidgetMinimized(!isBookingWidgetMinimized);
   };
 
   const toggleMenteesSection = () => {
@@ -638,151 +633,26 @@ export default function MentorPage() {
 
         {/* Right Main Content */}
         <div className="mentor-main-content">
-          {/* Dashboard Header */}
-          <div className="dashboard-header">
-            <div className="dashboard-title">
-              <h1>
-                {typeof currentUserProfile?.type === 'string' && currentUserProfile.type.toLowerCase() === 'mentee'
-                  ? 'Mentor Dashboard' 
-                  : 'Mentee Dashboard'
-                }
-              </h1>
-              <p>
-                {typeof currentUserProfile?.type === 'string' && currentUserProfile.type.toLowerCase() === 'mentee'
-                  ? 'Connect with experienced professionals who can guide you on your journey to success'
-                  : 'Find mentees who can benefit from your knowledge and experience'
-                }
-              </p>
-            </div>
-            
-            {/* Generated Profiles Info */}
-            <div className="generated-profiles-info">
-              <p className="info-text">
-                💡 <strong>Tip:</strong> This platform includes both real user profiles and generated test profiles 
-                (marked with 🎲) to help you explore the matching system and test features.
-              </p>
-              <p className="info-text" style={{ marginTop: '8px', fontSize: '0.9rem', opacity: '0.9' }}>
-                🎯 <strong>Smart Matching:</strong> Your matches are automatically calculated and ranked by compatibility percentage.
-              </p>
-            </div>
-          </div>
-
-          {/* User Profile Summary Cards */}
+          {/* Render appropriate dashboard based on user type */}
           {currentUserProfile && (
-            <div className="profile-summary-cards">
-              <div className="profile-card">
-                <div className="profile-card-header">
-                  <div className="profile-info">
-                    <div className={`profile-role ${typeof currentUserProfile.type === 'string' && currentUserProfile.type.toLowerCase() === 'mentor' ? 'mentor' : 'mentee'}`}>
-                      {typeof currentUserProfile.type === 'string' && currentUserProfile.type.toLowerCase() === 'mentor' ? 'Mentor' : 'Mentee'}
-                    </div>
-                    <div className="profile-name">
-                      {currentUserProfile.firstName} {currentUserProfile.lastName}
-                    </div>
-                  </div>
-                  <div className="profile-card-actions">
-                    <button 
-                      className="profile-edit-btn"
-                      onClick={handleProfileEdit}
-                      data-tooltip="Edit your profile information"
-                    >
-                      Edit Profile
-                    </button>
-                    <button 
-                      className="expand-toggle-btn"
-                      onClick={() => setIsProfileCardExpanded(!isProfileCardExpanded)}
-                      title={isProfileCardExpanded ? "Collapse profile details" : "Expand profile details"}
-                    >
-                      {isProfileCardExpanded ? '▼' : '▶'}
-                    </button>
-                  </div>
-                </div>
-                
-                {/* Expandable Profile Content */}
-                {isProfileCardExpanded && (
-                  <div className="profile-card-content">
-                    <div className="profile-details">
-                      <div className="detail-item">
-                        <span className="detail-label">Email:</span>
-                        <span className="detail-value">{currentUserProfile.email || 'Not provided'}</span>
-                      </div>
-                      <div className="detail-item">
-                        <span className="detail-label">Industry:</span>
-                        <span className="detail-value">{currentUserProfile.industry || 'Not specified'}</span>
-                      </div>
-                      <div className="detail-item">
-                        <span className="detail-label">Experience:</span>
-                        <span className="detail-value">{currentUserProfile.yearsOfExperience || 'Not specified'} years</span>
-                      </div>
-                      <div className="detail-item">
-                        <span className="detail-label">Skills:</span>
-                        <span className="detail-value">
-                          {currentUserProfile.skills && currentUserProfile.skills.length > 0 
-                            ? currentUserProfile.skills.slice(0, 3).join(', ') + (currentUserProfile.skills.length > 3 ? '...' : '')
-                            : 'No skills listed'
-                          }
-                        </span>
-                      </div>
-                    </div>
-                  </div>
-                )}
-              </div>
-
-              {typeof currentUserProfile?.type === 'string' && currentUserProfile.type.toLowerCase() === 'mentor' && (
-                <div className="profile-card">
-                  <div className="profile-card-header">
-                    <div className="profile-info">
-                      <div className="profile-role mentor">Mentor</div>
-                      <div className="profile-name">Availability Management</div>
-                    </div>
-                    <div className="profile-card-actions">
-                      <button 
-                        className="availability-manage-btn"
-                        onClick={handleAvailabilityManage}
-                        data-tooltip="Manage your availability schedule"
-                      >
-                        Manage Availability
-                      </button>
-                      <button 
-                        className="expand-toggle-btn"
-                        onClick={() => setIsAvailabilityCardExpanded(!isAvailabilityCardExpanded)}
-                        title={isAvailabilityCardExpanded ? "Collapse availability details" : "Expand availability details"}
-                      >
-                        {isAvailabilityCardExpanded ? '▼' : '▶'}
-                      </button>
-                    </div>
-                  </div>
-                  
-                  {/* Expandable Availability Content */}
-                  {isAvailabilityCardExpanded && (
-                    <div className="profile-card-content">
-                      <div className="availability-summary">
-                        <div className="availability-stat">
-                          <span className="stat-number">
-                            {Object.keys(mentorAvailability).length}
-                          </span>
-                          <span className="stat-label">Available Days</span>
-                        </div>
-                        <div className="availability-stat">
-                          <span className="stat-number">
-                            {Object.values(mentorAvailability).flat().length}
-                          </span>
-                          <span className="stat-label">Time Slots</span>
-                        </div>
-                        <div className="availability-stat">
-                          <span className="stat-number">
-                            {Object.values(mentorBookings).flat().filter(booking => 
-                              new Date(booking.sessionDate) > new Date()
-                            ).length}
-                          </span>
-                          <span className="stat-label">Upcoming Sessions</span>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
+            typeof currentUserProfile.type === 'string' && currentUserProfile.type.toLowerCase() === 'mentor' ? (
+              <MentorDashboard
+                currentUserProfile={currentUserProfile}
+                mentorAvailability={mentorAvailability}
+                mentorBookings={mentorBookings}
+                onProfileEdit={handleProfileEdit}
+                onAvailabilityManage={handleAvailabilityManage}
+                onViewAllBookings={handleViewAllBookings}
+                onAcceptBooking={handleAcceptBooking}
+                onRejectBooking={handleRejectBooking}
+                onCancelBooking={handleCancelBooking}
+              />
+            ) : (
+              <MenteeDashboard
+                currentUserProfile={currentUserProfile}
+                onProfileEdit={handleProfileEdit}
+              />
+            )
           )}
 
           {/* Matches Section - Moved from sidebar to main content */}
@@ -808,108 +678,6 @@ export default function MentorPage() {
               />
             </div>
           ) : null}
-
-          {/* Bookings Widget - Only show for mentors */}
-          {typeof currentUserProfile?.type === 'string' && currentUserProfile.type.toLowerCase() === 'mentor' && (
-            <div className={`bookings-widget-section ${isBookingWidgetMinimized ? 'minimized' : ''}`}>
-              <div 
-                className="bookings-widget-header"
-                onClick={toggleBookingWidget}
-                style={{ cursor: 'pointer' }}
-              >
-                <div className="bookings-widget-title">
-                  <h2>Your Bookings</h2>
-                  <button 
-                    className="minimize-toggle-btn"
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      toggleBookingWidget();
-                    }}
-                    title={isBookingWidgetMinimized ? "Expand widget" : "Minimize widget"}
-                  >
-                    {isBookingWidgetMinimized ? '▼' : '▲'}
-                  </button>
-                </div>
-                {!isBookingWidgetMinimized && (
-                  <p>Manage and view all your scheduled sessions</p>
-                )}
-              </div>
-              
-              {!isBookingWidgetMinimized && (
-                <div className="bookings-widget-content">
-                  <div className="bookings-summary">
-                    <div className="booking-stat">
-                      <span className="stat-number">
-                        {Object.values(mentorBookings).flat().length}
-                      </span>
-                      <span className="stat-label">Total Bookings</span>
-                    </div>
-                    <div className="booking-stat">
-                      <span className="stat-number">
-                        {Object.values(mentorBookings).flat().filter(booking => 
-                          new Date(booking.sessionDate) > new Date()
-                        ).length}
-                      </span>
-                      <span className="stat-label">Upcoming</span>
-                    </div>
-                    <div className="booking-stat">
-                      <span className="stat-number">
-                        {Object.values(mentorBookings).flat().filter(booking => 
-                          new Date(booking.sessionDate) <= new Date()
-                        ).length}
-                      </span>
-                      <span className="stat-label">Completed</span>
-                    </div>
-                  </div>
-                  
-                  <div className="bookings-list">
-                    {Object.values(mentorBookings).flat().length > 0 ? (
-                      Object.values(mentorBookings)
-                        .flat()
-                        .sort((a, b) => new Date(a.sessionDate).getTime() - new Date(b.sessionDate).getTime())
-                        .slice(0, 5)
-                        .map((booking, index) => (
-                          <div key={index} className="booking-item">
-                            <div className="booking-info">
-                              <div className="booking-header">
-                                <h4>{booking.menteeName || 'Mentee'}</h4>
-                                <span className={`booking-status ${booking.status || 'pending'}`}>
-                                  {booking.status || 'Pending'}
-                                </span>
-                              </div>
-                              <div className="booking-details">
-                                <span className="booking-date">
-                                  {new Date(booking.sessionDate).toLocaleDateString()}
-                                </span>
-                                <span className="booking-time">
-                                  {booking.startTime || 'Time TBD'}
-                                </span>
-                              </div>
-                            </div>
-                            <div className="booking-actions">
-                              <button className="booking-action-btn view">View Details</button>
-                              {booking.status === 'pending' && (
-                                <button className="booking-action-btn accept">Accept</button>
-                              )}
-                            </div>
-                          </div>
-                        ))
-                    ) : (
-                      <div className="no-bookings">
-                        <p>No bookings yet. Your scheduled sessions will appear here.</p>
-                      </div>
-                    )}
-                  </div>
-                  
-                  <div className="bookings-widget-footer">
-                    <button className="view-all-bookings-btn" onClick={handleViewAllBookings}>
-                      View All Bookings
-                    </button>
-                  </div>
-                </div>
-              )}
-            </div>
-          )}
 
           {/* Mentor Results */}
           <div className={`mentor-results ${isMenteesSectionMinimized ? 'minimized' : ''}`}>
